+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - New CSS handle `specificationItemProperty` plus the name of the specification.
 
+## [3.115.3] - 2020-06-02
+### Fixed
+- Brand image link changed to brand slug instead of brand name.
+
+## [3.115.2] - 2020-05-29
+
+### Fixed
+
+- Revert release v3.110.1 that was making images to not load in Chrome 83.
+
 ## [3.115.1] - 2020-05-28
 ### Fixed